Add lowercase when adding a letter to list_of_guesses.
Prior to this change, the code was incorrectly adding both uppercase and
lowercase letters to list_of_guesses, and therefore not showing that the letter
had already been tried.

This change converts letters to lowercase when adding to list_of_guesses
and also before checking if a letter is in list_of_guesses.